Move data structs to own file
